Proofs completed
